refactor(events): split EventActionSidebar under the line cap (#544)
Extract the staff/owner "Manage Event" section (admin links +
announcement modal + its events/tiers queries) into
EventManageSection.svelte. Behavior-preserving move only.
EventActionSidebar 761 -> 601 lines.
